cleanup "warning: deprecated conversion from string constant to 'char*'"
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@4437 640d5228-2204-0410-8d71-8f46fa6850b4

+++ xpms2d/class/DataFile.h
@@ -124,8 +124,8 @@ protected:
   HeaderType	_fileHeaderType;
 
   Index		*indices;
-  long		currPhys;
-  int		currLR;
+  size_t	currPhys;
+  size_t	currLR;
   size_t	nIndices;
 
   P2d_rec	physRecord[P2DLRPR], *testRecP;
